requirements.txt
pandas
numpy
pillow
pygame
opencv_python
tensorflow
ultralytics